# VESPA/ves-simu/assets/open-api-spec.yaml
openapi: "3.0.1"
info:
  title: VES-Collector simulator
  description: VES-Collector simulator API
  version: "7.1"
...
paths:
  /eventListener/v7/:
    summary: VES-Collector event API
    post:
      summary: Publish a single event
      description: |
        This is the official VES-Collector API for posting events.
        The simulator aims at simulating it
      tags: [VES]
      requestBody:
        description: The body contains the event to be published
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event:
                  $ref: "#/components/schemas/event"
            example: {
              "event": {
                "commonEventHeader": {
                  "domain": "measurementsForVfScaling",
                  "eventId": "12",
                  "eventName": "Mfvs_foobar",
                  "lastEpochMicrosec": 1536754800000000,
                  "priority": "Normal",
                  "reportingEntityName": "test-pro-1",
                  "sequence": 0,
                  "sourceName": "test-pro-1",
                  "startEpochMicrosec": 1536754500000000,
                  "version": 3
                },
                "measurementsForVfScalingFields": {
                  "cpuUsageArray": [
                    {
                      "cpuIdentifier": "0",
                      "percentUsage": 1.5666666666530205
                    },
                    {
                      "cpuIdentifier": "1",
                      "percentUsage": 1.2833333333468886
                    }
                  ],
                  "measurementInterval": 300,
                  "measurementsForVfScalingVersion": 2,
                  "memoryUsageArray": [
                    {
                      "memoryFree": 1373691904,
                      "memoryUsed": 2600394752,
                      "vmIdentifier": "test-pro-1"
                    }
                  ]
                }
              }
            }
      responses:
        202:
          description: The request has been accepted for processing
        400:
          description: Many possible reasons not specified by the other codes (e.g., missing required parameters or incorrect format). The response body may include a further exception code and text. HTTP 400 errors may be mapped to SVC0001 (general service error), SVC0002 (bad parameter), SVC2000 (general service error with details) or PO9003 (message content size exceeds the allowable limit).
        401:
          description: Authentication failed or was not provided. HTTP 401 errors may be mapped to POL0001 (general policy error) or POL2000 (general policy error with details).
        404:
          description: The server has not found anything matching the Request-URI. No indication is given of whether the condition is temporary or permanent.
        405:
          description: A request was made of a resource using a request method not supported by that resource (e.g., using PUT on a REST resource that only supports POST).
        500:
          description: The server encountered an internal error or timed out; please retry (general catch-all server-side error).HTTP 500 errors may be mapped to SVC1000 (no server resources).
  /eventListener/v7/eventBatch:
    summary: VES-Collector batch API
    post:
      summary: Publish a batch of multiple events
      description: |
        This is the official VES-Collector API for posting batch of events.
        The simulator aims at simulating it
      tags: [VES]
      requestBody:
        description: The body contains the list of events to be published
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventList:
                  $ref: "#/components/schemas/eventList"
            example: {
              "eventList": [{
                "commonEventHeader": {
                  "domain": "measurementsForVfScaling",
                  "eventId": "12",
                  "eventName": "Mfvs_foobar",
                  "lastEpochMicrosec": 1536754800000000,
                  "priority": "Normal",
                  "reportingEntityName": "test-pro-1",
                  "sequence": 0,
                  "sourceName": "test-pro-1",
                  "startEpochMicrosec": 1536754500000000,
                  "version": 3
                },
                "measurementsForVfScalingFields": {
                  "cpuUsageArray": [
                    {
                      "cpuIdentifier": "0",
                      "percentUsage": 1.5666666666530205
                    },
                    {
                      "cpuIdentifier": "1",
                      "percentUsage": 1.2833333333468886
                    }
                  ],
                  "measurementInterval": 300,
                  "measurementsForVfScalingVersion": 2,
                  "memoryUsageArray": [
                    {
                      "memoryFree": 1373691904,
                      "memoryUsed": 2600394752,
                      "vmIdentifier": "test-pro-1"
                    }
                  ]
                }
              }]
            }
      responses:
        202:
          description: The request has been accepted for processing
        400:
          description: Many possible reasons not specified by the other codes (e.g., missing required parameters or incorrect format). The response body may include a further exception code and text. HTTP 400 errors may be mapped to SVC0001 (general service error), SVC0002 (bad parameter), SVC2000 (general service error with details) or PO9003 (message content size exceeds the allowable limit).
        401:
          description: Authentication failed or was not provided. HTTP 401 errors may be mapped to POL0001 (general policy error) or POL2000 (general policy error with details).
        404:
          description: The server has not found anything matching the Request-URI. No indication is given of whether the condition is temporary or permanent.
        405:
          description: A request was made of a resource using a request method not supported by that resource (e.g., using PUT on a REST resource that only supports POST).
        500:
          description: The server encountered an internal error or timed out; please retry (general catch-all server-side error).HTTP 500 errors may be mapped to SVC1000 (no server resources).
  /testControl/v7/events:
    summary: Manipulate received events
    get: 
      summary: List received events
      description: Will return all the events received and stored by the server
      tags: [Simulator]
      parameters:
        - name: domain
          schema:
            type: string
          in: query
          description: An optional VES event domain to filter on
          required: false
        - name: entity
          schema:
            type: string
          in: query
          description: An optional VES event entity to filter on
          required: false
        - name: clear
          schema:
            type: integer
            enum: [0, 1]
          in: query
          description: Clear saved events after a successfull query
          required: false

      responses:
        200:
          description: Events successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/event"
    delete:
      summary: Delete stored events
      description: Delete from memory all the stored events
      tags: [Simulator]
      responses:
        200:
          description: Events deleted properly
  /testControl/v7/commandList:
    summary: Control commands to send to VES-Agent
    post:
      summary: Inject commands
      description: Update commands to send to VES-Agent
      tags: [Simulator]
      requestBody:
        description: Request body contains a structure with list of commands to inject
        required: true
        content:
          application/json:
            schema:
              properties:
                commandList:
                  $ref: "#/components/schemas/commandList"
            example:
              commandList:
                - commandType:  heartbeatIntervalChange
                  heartbeatInterval: 30
      responses:
        200:
          description: Commands successfully set
        400:
          description: Something wrong happened
...
components:
...
  "schemas": {
...
    "command": {
      "description": "command from an event collector toward an event source",
      "type": "object",
      "properties": {
        "commandType": {
          "type": "string",
          "enum": [
            "heartbeatIntervalChange",
            "measurementIntervalChange",
            "provideThrottlingState",
            "throttlingSpecification"
          ]
        },
        "eventDomainThrottleSpecification": {
          "$ref": "#/components/schemas/eventDomainThrottleSpecification"
        },
        "heartbeatInterval": {
          "type": "integer"
        },
        "measurementInterval": {
          "type": "integer"
        }
      },
      "required": [
        "commandType"
      ]
    },
    "commandList": {
      "description": "array of commands from an event collector toward an event source",
      "type": "array",
      "items": {
        "$ref": "#/components/schemas/command"
      },
      "minItems": 0
    },
    "commonEventHeader": {
      "description": "fields common to all events",
      "type": "object",
      "properties": {
        "domain": {
          "description": "the eventing domain associated with the event",
          "type": "string",
          "enum": [
            "fault",
            "heartbeat",
            "measurementsForVfScaling",
            "mobileFlow",
            "other",
            "sipSignaling",
            "stateChange",
            "syslog",
            "thresholdCrossingAlert",
            "voiceQuality"
          ]
        },
        "eventId": {
          "description": "event key that is unique to the event source",
          "type": "string"
        },
        "eventName": {
          "description": "unique event name",
          "type": "string"
        },
        "eventType": {
          "description": "for example - applicationVnf, guestOS, hostOS, platform",
          "type": "string"
        },
        "internalHeaderFields": {
          "$ref": "#/components/schemas/internalHeaderFields"
        },
        "lastEpochMicrosec": {
          "description": "the latest unix time aka epoch time associated with the event from any component--as microseconds elapsed since 1 Jan 1970 not including leap seconds",
          "type": "number"
        },
        "nfcNamingCode": {
          "description": "3 character network function component type, aligned with vfc naming standards",
          "type": "string"
        },
        "nfNamingCode": {
          "description": "4 character network function type, aligned with vnf naming standards",
          "type": "string"
        },
        "priority": {
          "description": "processing priority",
          "type": "string",
          "enum": [
            "High",
            "Medium",
            "Normal",
            "Low"
          ]
        },
        "reportingEntityId": {
          "description": "UUID identifying the entity reporting the event, for example an OAM VM; must be populated by the ATT enrichment process",
          "type": "string"
        },
        "reportingEntityName": {
          "description": "name of the entity reporting the event, for example, an EMS name; may be the same as sourceName",
          "type": "string"
        },
        "sequence": {
          "description": "ordering of events communicated by an event source instance or 0 if not needed",
          "type": "integer"
        },
        "sourceId": {
          "description": "UUID identifying the entity experiencing the event issue; must be populated by the ATT enrichment process",
          "type": "string"
        },
        "sourceName": {
          "description": "name of the entity experiencing the event issue",
          "type": "string"
        },
        "startEpochMicrosec": {
          "description": "the earliest unix time aka epoch time associated with the event from any component--as microseconds elapsed since 1 Jan 1970 not including leap seconds",
          "type": "number"
        },
        "version": {
          "description": "version of the event header",
          "type": "number"
        }
      },
      "required": [
        "domain",
        "eventId",
        "eventName",
        "lastEpochMicrosec",
        "priority",
        "reportingEntityName",
        "sequence",
        "sourceName",
        "startEpochMicrosec",
        "version"
      ]
    },
...
    "event": {
      "description": "the root level of the common event format",
      "type": "object",
      "properties": {
        "commonEventHeader": {
          "$ref": "#/components/schemas/commonEventHeader"
        },
        "faultFields": {
          "$ref": "#/components/schemas/faultFields"
        },
        "heartbeatFields": {
          "$ref": "#/components/schemas/heartbeatFields"
        },
        "measurementsForVfScalingFields": {
          "$ref": "#/components/schemas/measurementsForVfScalingFields"
        },
        "mobileFlowFields": {
          "$ref": "#/components/schemas/mobileFlowFields"
        },
        "otherFields": {
          "$ref": "#/components/schemas/otherFields"
        },
        "sipSignalingFields": {
          "$ref": "#/components/schemas/sipSignalingFields"
        },
        "stateChangeFields": {
          "$ref": "#/components/schemas/stateChangeFields"
        },
        "syslogFields": {
          "$ref": "#/components/schemas/syslogFields"
        },
        "thresholdCrossingAlertFields": {
          "$ref": "#/components/schemas/thresholdCrossingAlertFields"
        },
        "voiceQualityFields": {
          "$ref": "#/components/schemas/voiceQualityFields"
        }
      },
      "required": [
        "commonEventHeader"
      ]
    },
...
    "eventList": {
      "description": "array of events",
      "type": "array",
      "items": {
        "$ref": "#/components/schemas/event"
      }
    },
...
    "faultFields": {
      "description": "fields specific to fault events",
      "type": "object",
      "properties": {
        "alarmAdditionalInformation": {
          "description": "additional alarm information",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "alarmCondition": {
          "description": "alarm condition reported by the device",
          "type": "string"
        },
        "alarmInterfaceA": {
          "description": "card, port, channel or interface name of the device generating the alarm",
          "type": "string"
        },
        "eventCategory": {
          "description": "Event category, for example: license, link, routing, security, signaling",
          "type": "string"
        },
        "eventSeverity": {
          "description": "event severity",
          "type": "string",
          "enum": [
            "CRITICAL",
            "MAJOR",
            "MINOR",
            "WARNING",
            "NORMAL"
          ]
        },
        "eventSourceType": {
          "description": "type of event source; examples: card, host, other, port, portThreshold, router, slotThreshold, switch, virtualMachine, virtualNetworkFunction",
          "type": "string"
        },
        "faultFieldsVersion": {
          "description": "version of the faultFields block",
          "type": "number"
        },
        "specificProblem": {
          "description": "short description of the alarm or problem",
          "type": "string"
        },
        "vfStatus": {
          "description": "virtual function status enumeration",
          "type": "string",
          "enum": [
            "Active",
            "Idle",
            "Preparing to terminate",
            "Ready to terminate",
            "Requesting termination"
          ]
        }
      },
      "required": [
        "alarmCondition",
        "eventSeverity",
        "eventSourceType",
        "faultFieldsVersion",
        "specificProblem",
        "vfStatus"
      ]
    },
...
    "heartbeatFields": {
      "description": "optional field block for fields specific to heartbeat events",
      "type": "object",
      "properties": {
        "additionalFields": {
          "description": "additional heartbeat fields if needed",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "heartbeatFieldsVersion": {
          "description": "version of the heartbeatFields block",
          "type": "number"
        },
        "heartbeatInterval": {
          "description": "current heartbeat interval in seconds",
          "type": "integer"
        }
      },
      "required": [
        "heartbeatFieldsVersion",
        "heartbeatInterval"
      ]
    },
...
    "measurementsForVfScalingFields": {
      "description": "measurementsForVfScaling fields",
      "type": "object",
      "properties": {
        "additionalFields": {
          "description": "additional name-value-pair fields",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "additionalMeasurements": {
          "description": "array of named name-value-pair arrays",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/namedArrayOfFields"
          }
        },
        "additionalObjects": {
          "description": "array of JSON objects described by name, schema and other meta-information",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/jsonObject"
          }
        },
        "codecUsageArray": {
          "description": "array of codecs in use",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/codecsInUse"
          }
        },
        "concurrentSessions": {
          "description": "peak concurrent sessions for the VM or VNF over the measurementInterval",
          "type": "integer"
        },
        "configuredEntities": {
          "description": "over the measurementInterval, peak total number of: users, subscribers, devices, adjacencies, etc., for the VM, or subscribers, devices, etc., for the VNF",
          "type": "integer"
        },
        "cpuUsageArray": {
          "description": "usage of an array of CPUs",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/cpuUsage"
          }
        },
        "diskUsageArray": {
          "description": "usage of an array of disks",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/diskUsage"
          }
        },
        "featureUsageArray": {
          "description": "array of features in use",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/featuresInUse"
          }
        },
        "filesystemUsageArray": {
          "description": "filesystem usage of the VM on which the VNFC reporting the event is running",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/filesystemUsage"
          }
        },
        "latencyDistribution": {
          "description": "array of integers representing counts of requests whose latency in milliseconds falls within per-VNF configured ranges",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/latencyBucketMeasure"
          }
        },
        "meanRequestLatency": {
          "description": "mean seconds required to respond to each request for the VM on which the VNFC reporting the event is running",
          "type": "number"
        },
        "measurementInterval": {
          "description": "interval over which measurements are being reported in seconds",
          "type": "number"
        },
        "measurementsForVfScalingVersion": {
          "description": "version of the measurementsForVfScaling block",
          "type": "number"
        },
        "memoryUsageArray": {
          "description": "memory usage of an array of VMs",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/memoryUsage"
          }
        },
        "numberOfMediaPortsInUse": {
          "description": "number of media ports in use",
          "type": "integer"
        },
        "requestRate": {
          "description": "peak rate of service requests per second to the VNF over the measurementInterval",
          "type": "number"
        },
        "vnfcScalingMetric": {
          "description": "represents busy-ness of the VNF from 0 to 100 as reported by the VNFC",
          "type": "integer"
        },
        "vNicPerformanceArray": {
          "description": "usage of an array of virtual network interface cards",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/vNicPerformance"
          }
        }
      },
      "required": [
        "measurementInterval",
        "measurementsForVfScalingVersion"
      ]
    },
...
    "mobileFlowFields": {
      "description": "mobileFlow fields",
      "type": "object",
      "properties": {
        "additionalFields": {
          "description": "additional mobileFlow fields if needed",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "applicationType": {
          "description": "Application type inferred",
          "type": "string"
        },
        "appProtocolType": {
          "description": "application protocol",
          "type": "string"
        },
        "appProtocolVersion": {
          "description": "application protocol version",
          "type": "string"
        },
        "cid": {
          "description": "cell id",
          "type": "string"
        },
        "connectionType": {
          "description": "Abbreviation referencing a 3GPP reference point e.g., S1-U, S11, etc",
          "type": "string"
        },
        "ecgi": {
          "description": "Evolved Cell Global Id",
          "type": "string"
        },
        "flowDirection": {
          "description": "Flow direction, indicating if the reporting node is the source of the flow or destination for the flow",
          "type": "string"
        },
        "gtpPerFlowMetrics": {
          "$ref": "#/components/schemas/gtpPerFlowMetrics"
        },
        "gtpProtocolType": {
          "description": "GTP protocol",
          "type": "string"
        },
        "gtpVersion": {
          "description": "GTP protocol version",
          "type": "string"
        },
        "httpHeader": {
          "description": "HTTP request header, if the flow connects to a node referenced by HTTP",
          "type": "string"
        },
        "imei": {
          "description": "IMEI for the subscriber UE used in this flow, if the flow connects to a mobile device",
          "type": "string"
        },
        "imsi": {
          "description": "IMSI for the subscriber UE used in this flow, if the flow connects to a mobile device",
          "type": "string"
        },
        "ipProtocolType": {
          "description": "IP protocol type e.g., TCP, UDP, RTP...",
          "type": "string"
        },
        "ipVersion": {
          "description": "IP protocol version e.g., IPv4, IPv6",
          "type": "string"
        },
        "lac": {
          "description": "location area code",
          "type": "string"
        },
        "mcc": {
          "description": "mobile country code",
          "type": "string"
        },
        "mnc": {
          "description": "mobile network code",
          "type": "string"
        },
        "mobileFlowFieldsVersion": {
          "description": "version of the mobileFlowFields block",
          "type": "number"
        },
        "msisdn": {
          "description": "MSISDN for the subscriber UE used in this flow, as an integer, if the flow connects to a mobile device",
          "type": "string"
        },
        "otherEndpointIpAddress": {
          "description": "IP address for the other endpoint, as used for the flow being reported on",
          "type": "string"
        },
        "otherEndpointPort": {
          "description": "IP Port for the reporting entity, as used for the flow being reported on",
          "type": "integer"
        },
        "otherFunctionalRole": {
          "description": "Functional role of the other endpoint for the flow being reported on e.g., MME, S-GW, P-GW, PCRF...",
          "type": "string"
        },
        "rac": {
          "description": "routing area code",
          "type": "string"
        },
        "radioAccessTechnology": {
          "description": "Radio Access Technology e.g., 2G, 3G, LTE",
          "type": "string"
        },
        "reportingEndpointIpAddr": {
          "description": "IP address for the reporting entity, as used for the flow being reported on",
          "type": "string"
        },
        "reportingEndpointPort": {
          "description": "IP port for the reporting entity, as used for the flow being reported on",
          "type": "integer"
        },
        "sac": {
          "description": "service area code",
          "type": "string"
        },
        "samplingAlgorithm": {
          "description": "Integer identifier for the sampling algorithm or rule being applied in calculating the flow metrics if metrics are calculated based on a sample of packets, or 0 if no sampling is applied",
          "type": "integer"
        },
        "tac": {
          "description": "transport area code",
          "type": "string"
        },
        "tunnelId": {
          "description": "tunnel identifier",
          "type": "string"
        },
        "vlanId": {
          "description": "VLAN identifier used by this flow",
          "type": "string"
        }
      },
      "required": [
        "flowDirection",
        "gtpPerFlowMetrics",
        "ipProtocolType",
        "ipVersion",
        "mobileFlowFieldsVersion",
        "otherEndpointIpAddress",
        "otherEndpointPort",
        "reportingEndpointIpAddr",
        "reportingEndpointPort"
      ]
    },
...
    "otherFields": {
      "description": "fields for events belonging to the 'other' domain of the commonEventHeader domain enumeration",
      "type": "object",
      "properties": {
        "hashOfNameValuePairArrays": {
          "description": "array of named name-value-pair arrays",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/namedArrayOfFields"
          }
        },
        "jsonObjects": {
          "description": "array of JSON objects described by name, schema and other meta-information",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/jsonObject"
          }
        },
        "nameValuePairs": {
          "description": "array of name-value pairs",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "otherFieldsVersion": {
          "description": "version of the otherFields block",
          "type": "number"
        }
      },
      "required": [
        "otherFieldsVersion"
      ]
    },
...
    "sipSignalingFields": {
      "description": "sip signaling fields",
      "type": "object",
      "properties": {
        "additionalInformation": {
          "description": "additional sip signaling fields if needed",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "compressedSip": {
          "description": "the full SIP request/response including headers and bodies",
          "type": "string"
        },
        "correlator": {
          "description": "this is the same for all events on this call",
          "type": "string"
        },
        "localIpAddress": {
          "description": "IP address on VNF",
          "type": "string"
        },
        "localPort": {
          "description": "port on VNF",
          "type": "string"
        },
        "remoteIpAddress": {
          "description": "IP address of peer endpoint",
          "type": "string"
        },
        "remotePort": {
          "description": "port of peer endpoint",
          "type": "string"
        },
        "sipSignalingFieldsVersion": {
          "description": "version of the sipSignalingFields block",
          "type": "number"
        },
        "summarySip": {
          "description": "the SIP Method or Response (‘INVITE’, ‘200 OK’, ‘BYE’, etc)",
          "type": "string"
        },
        "vendorVnfNameFields": {
          "$ref": "#/components/schemas/vendorVnfNameFields"
        }
      },
      "required": [
        "correlator",
        "localIpAddress",
        "localPort",
        "remoteIpAddress",
        "remotePort",
        "sipSignalingFieldsVersion",
        "vendorVnfNameFields"
      ]
    },
    "stateChangeFields": {
      "description": "stateChange fields",
      "type": "object",
      "properties": {
        "additionalFields": {
          "description": "additional stateChange fields if needed",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "newState": {
          "description": "new state of the entity",
          "type": "string",
          "enum": [
            "inService",
            "maintenance",
            "outOfService"
          ]
        },
        "oldState": {
          "description": "previous state of the entity",
          "type": "string",
          "enum": [
            "inService",
            "maintenance",
            "outOfService"
          ]
        },
        "stateChangeFieldsVersion": {
          "description": "version of the stateChangeFields block",
          "type": "number"
        },
        "stateInterface": {
          "description": "card or port name of the entity that changed state",
          "type": "string"
        }
      },
      "required": [
        "newState",
        "oldState",
        "stateChangeFieldsVersion",
        "stateInterface"
      ]
    },
...
    "voiceQualityFields": {
      "description": "provides statistics related to customer facing voice products",
      "type": "object",
      "properties": {
        "additionalInformation": {
          "description": "additional voice quality fields if needed",
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/field"
          }
        },
        "calleeSideCodec": {
          "description": "callee codec for the call",
          "type": "string"
        },
        "callerSideCodec": {
          "description": "caller codec for the call",
          "type": "string"
        },
        "correlator": {
          "description": "this is the same for all events on this call",
          "type": "string"
        },
        "endOfCallVqmSummaries": {
          "$ref": "#/components/schemas/endOfCallVqmSummaries"
        },
        "phoneNumber": {
          "description": "phone number associated with the correlator",
          "type": "string"
        },
        "midCallRtcp": {
          "description": "Base64 encoding of the binary RTCP data excluding Eth/IP/UDP headers",
          "type": "string"
        },
        "vendorVnfNameFields": {
          "$ref": "#/components/schemas/vendorVnfNameFields"
        },
        "voiceQualityFieldsVersion": {
          "description": "version of the voiceQualityFields block",
          "type": "number"
        }
      },
      "required": [
        "calleeSideCodec",
        "callerSideCodec",
        "correlator",
        "midCallRtcp",
        "vendorVnfNameFields",
        "voiceQualityFieldsVersion"
      ]
    }
...